Fix a bug where the trailers of a streaming response was dropped by RetryingClient (#6213)
Motivation:

During the implementation of #5714, `HttpResponse.split()` was used to
extract the headers from a streaming response. After applying a
`RetryRule`, the new response was reconstructed in

https://github.com/line/armeria/blob/3a35abe68c80c0af8594600ce677df49e09a9e0e/core/src/main/java/com/linecorp/armeria/client/retry/RetryingClient.java#L429
However, because `SplitHttpResponse.body()` only publishes the data
payloads and handles trailers separately, the trailers was missing in
the reconstructed response, leading to a bug.

Modifications:

- Fixed `SurroundingPublisher` to allow the last element to be emitted
asynchronously.
- Added `SplitHttpRequest.unsplit()` and `SplitHttpResponse.unsplit()`
to re-create an `HttpResponse` from the split headers, body and
trailers.
- Fixed `RetryingClient` to use `.unsplit()` for deliveriong trailers
correctly.

Result:

Fix a regression where RetryingClient dropped trailers of streaming
responses (since 1.32.4)