Server doesn't send a `RST_STREAM` frame after `endStream` is sent (#6279)
Motivation:

The PR #6166 modified server-side behavior to send a `RST_STREAM` frame
when the corresponding Armeria's `HttpRequest` and `HttpResponse` are
completed. The assumption was that since the user was done using the
`HttpRequest` and `HttpResponse`, the underlying stream can be cleaned
up.


https://github.com/line/armeria/blob/a4bcb3e3b2a90a41b0066601d97da2fb04af5347/core/src/main/java/com/linecorp/armeria/server/ServerHttp2ObjectEncoder.java#L146-L149

The code was written to check Netty's `Http2Stream` before sending a
`RST_STREAM` - this is because sending a `endStream` would close the
underlying stream anyways regardless of whether a `RST_STREAM` is sent.

However, when frames are flow controlled, `Http2Stream`'s state is
updated after armeria's post-response future is invoked. To be more
specific, it seems like Netty's implementation updates the state by
adding a listener on the `write*` future when the flow controlled frame
is completed.
This is awkward because Armeria also completes the `HttpResponse` when a
write future with `endStream` is completed.

1. Armeria layer: Write `endStream` and wait on the write future before
invoking `res.close`


https://github.com/line/armeria/blob/a4bcb3e3b2a90a41b0066601d97da2fb04af5347/core/src/main/java/com/linecorp/armeria/server/AbstractHttpResponseSubscriber.java#L397-L417

2. `RequestAndResponseCompleteHandler` is completed


https://github.com/line/armeria/blob/a4bcb3e3b2a90a41b0066601d97da2fb04af5347/core/src/main/java/com/linecorp/armeria/server/HttpServerHandler.java#L847-L848

4. `Http2Stream.state` is updated (the future is invoked after
Armeria's)


https://github.com/netty/netty/blob/a919dd3a53a50da080dc925a6fb085c81cab6294/codec-http2/src/main/java/io/netty/handler/codec/http2/Http2ConnectionHandler.java#L629

Note that there is no issue when receiving requests since the state is
updated before the `Http2FrameListener` is invoked.

i.e.
https://github.com/netty/netty/blob/a919dd3a53a50da080dc925a6fb085c81cab6294/codec-http2/src/main/java/io/netty/handler/codec/http2/DefaultHttp2ConnectionDecoder.java#L257

I propose that we track a separate state which determines whether an
`endStream` was sent. This value can be used to determine whether a
separate `RST_STREAM` should be sent to close the underlying stream.

Modifications:

- Added a `endStreamSentKey` property key which indicates whether an
`endStream` was sent for a `Http2Stream`.
- Check if `endStreamSentKey` was set when determining whether
`RST_STREAM` should be used to clean up the underlying stream.
- To determine if the local side (server->client) is open, both
`Http2Stream.State.localSideOpen() == true` and `endStreamSentKey ==
false` must be satisfied.
- If the remote is open (client->server), then the `RST_STREAM` is sent
regardless.
- For client-side, a late `RST_STREAM` is always logged if the response
is closed. If the stream is active, then the `RST_STREAM` was sent to
close the stream and hence there is no need to log with WARN level.
- Added `SimpleHttp2Connection` to easily send and receive HTTP2 frames
directly from test code

Result:

- An unecessary `RST_STREAM` is not sent after the server sends an
`endStream`.

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->