Introduce `XdsClusterManager` to cache `Cluster`s (#6261)
Motivation:

This changeset can be thought of as a refactoring for the next PR
https://github.com/line/armeria/pull/6096.
Functionally, there should be no changes.

##### Cluster caching

Currently, a
[ClusterManager](https://github.com/line/armeria/blob/e766e37f42b46427ac2c443ff4aa5cabea12c965/xds/src/main/java/com/linecorp/armeria/xds/client/endpoint/ClusterManager.java)
is created every time a new `XdsEndpointGroup` is created.

While this was easy to implement at the time, this poses a problem where
each `XdsEndpointGroup` creates a backing `EndpointGroup` such as
`HealthCheckedEndpointGroup` or `DnsAddressEndpointGroup`. Moreover,
even having multiple clusters contained in different routes would also
consume resources. (e.g. `route1` with `cluster1` and `route2` with
`cluster1`. Although `cluster1` is the same, they are contained in a
different `RouteSnapshot` and thus handled twice)

A cluster is expected to be uniquely identified by its name as specified
in the
[documentation](https://www.envoyproxy.io/docs/envoy/latest/api-v3/config/cluster/v3/cluster.proto#envoy-v3-api-msg-config-cluster-v3-cluster),
and this is reflected in the upstream implementation as well.
[ref](https://github.com/envoyproxy/envoy/blob/95be4128e61406ea8786766dc7a738ffaf7ade3b/source/common/upstream/cluster_manager_impl.h#L971)

I propose that for each `XdsBootstrap`, clusters and the resources
attached to them are cached by name. The centralized location for
managing these clusters is named `XdsClusterManager` after
[ClusterManager](https://www.envoyproxy.io/docs/envoy/latest/intro/arch_overview/upstream/cluster_manager).

In the future, `XdsClusterManager` will take over the responsibility of
the current `ClusterManager` and also manage resources (health checking,
dns, etc..) associated with clusters. While caching resources isn't done
in this PR,
the following [POC PR](https://github.com/line/armeria/pull/6096) can be
checked for reference on how this will be done.

##### `XdsClusterManager` initialization

The changes in this PR also try to align closer to the upstream
implementation.

There is an inherent dependency between adding clusters and creating a
bootstrap. Bootstrap clusters should be created on initialization; yet
bootstrap clusters may require connection to a control plane (which may
require other bootstraps). In order to resolve this, a couple of
refactoring efforts have taken place.

The initialization process of `XdsClusterManager` has been modified to
closely follows that of upstream. This guarantees that bootstrapped
clusters will initialize correctly irregardless of the declaration order
and matches upstream behavior.
([ref](https://github.com/envoyproxy/envoy/blob/95be4128e61406ea8786766dc7a738ffaf7ade3b/source/common/upstream/cluster_manager_impl.cc#L367)).
Essentially, start up is done in two phases 1) non-eds cluster
initialization 2) eds cluster initialization. This is done since it is
possible that eds clusters require a bootstrap to connect to a control
plane (relying on non-eds clusters).

Another implication of the change above is that a bootstrap is not
created when bootstrap clusters are initialized. In order to resolve
this circular dependency, the logic where remote subscription (to the
control plane) takes place has been abstracted to a
`ControlPlaneClientManager`. Now that subscription doesn't require a
bootstrap and requires a `ControlPlaneClientManager`, a new
`SubscriptionContext` has been introduced to allow for easier
subscription. Since non-eds bootstrap clusters don't require connection
to a control plane, a separate `StaticSubscriptionContext` is used.

Modifications:

- `Cluster` caching related changes
- Primer related logic has been removed and replaced with a throw-away
watcher based approach. The reasoning is that primers are not reliable
anymore given that `Cluster`s are now cached.
- `ResourceNode`s now support multiple snapshot watchers. This change is
made so that `XdsClusterManager` can determine if a cluster no longer
has to be cached depending on the watchers.
- `XdsClusterManager` initialization related changes
- `ControlPlaneClientManager` is introduced which manages
`ConfigSourceClient`s
- `SubscriptionContext` is introduced to remove the circular dependency
between `ResourceNode`s and bootstrap.
  - `BootstrapClusters` is now initialized in two phases
- Misc) Creating a `ClusterRoot` or `ListenerRoot` can cause logic in
the `ClusterManager` to be executed from a non-eventloop. Modified the
logic so that user threads cannot modify `ClusterManager` or the
internal bootstrap state.
- Misc) `ClusterXdsResource#upstreamTlsContext` is now created once and
cached

Result:

- Preparation for introducing xDS http filter support at
https://github.com/line/armeria/pull/6096

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->