Store method descriptor to the client request context for GRPC calls - take 2 (#6300)
Motivation:

I want the method descriptor to be available to the client decorators and/or retry logic to be able to read method options (custom or not) in the logic.
WIthout this we would need to rely on converting from the url to the method name and manually construct a descriptor repo to be able to get that information. 

Modifications:

I made a class similar with private and public variants to hold both CallOptions and the MethodDescriptor inside the request context. 

Result:

There are no changes in the current behavior but more information is available for custom logic.

The new logic can be used like this:
```java
MyGrpcStub client = GrpcClients
    .builder(grpcServerUri)
       .decorator((delegate, ctx, req) -> {
          MethodDescriptor descriptor = GrpcClientCall.getMethodDescriptor(ctx);
           boolean safe = descriptor.isIdempotent() || descriptor.isSafe()
           ...
```

This is an alternative implementation to #6291 

<!--
Visit this URL to learn more about how to write a pull request description:
https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->