(Thrift Field Masker Pt2) Implement field masking for thrift (#6311)
Motivation:

Following field maskers for annotated services at #6232 , this PR
attempts to introduce `FieldMasker` support for thrift.

To ensure that deserialization is also supported natively using
`TProtocol`, the design of thrift uses a variant of `TProtocol` which
keeps track of the field of a specified instance.
Unmasking support is to be done in the ensuing PR for easier reviews.

Modifications:

- Introduce `RpcRequestSerializer`, `RpcResponseSerializer` to serialize
rpc requests
- `ThriftFieldInfo`, `ThriftFieldMaskerSelector`, and
`ThriftFieldMaskerSelectorProvider` are implemented to support the
`FieldMasker` API.
- `TBaseSelectorCache`, `TBaseSerializer`, `MaskingTProtocol` are
implemented to actually apply field masking during serialization
- Added the ability to specify the `annotations_as_metadata` option to
the gradle scripts

Result:

- Users can use the `FieldMasker` API for thrift services

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->