feat: accept empty content body in HttpJsonTranscodingService (#6325)
Motivation

HttpJsonTranscodingService currently fails with 400 Bad Request when handling a request with Content-Type: application/json and no body, even if the corresponding google.api.http rule uses body: "*". This behavior is inconsistent with common implementations such as Envoy and grpc-gateway, which treat empty bodies as {} in such cases.

⸻

Modifications

- **HttpJsonTranscodingService**
  - Added a fallback to use an empty JSON object when body empty ( content-length: 0 )
  - Treat **empty request bodies** as `{}` when **all** of the following are true:
    - **Content-Type**: `application/json`
    - **google.api.http rule**: `body: "*"`

⸻

Result

- Closes #6319.
- Users can now send requests with Content-Type: application/json and an empty body without receiving a 400 Bad Request error.
- Ensures better compatibility with gRPC transcoding conventions from other ecosystems