Implement xDS virtual host routing (#6322)
Motivation:

This changeset attempts to implement routing based on virtual hosts.

The relevant documentation regarding matching can be found in the
following locations:
-
https://www.envoyproxy.io/docs/envoy/latest/api-v3/config/route/v3/route.proto
-
https://www.envoyproxy.io/docs/envoy/latest/api-v3/config/route/v3/route_components.proto#envoy-v3-api-msg-config-route-v3-virtualhost

The rules for domain matching can be found in detail at the
`VirtualHost#domains` section.

The relevant upstream implementation can be found in the router
configuration:
-
https://github.com/envoyproxy/envoy/blob/0d341b982a24f3ddb9f4fd3a9bdd8880561103be/source/common/router/config_impl.cc#L1930-L1976

Note that there is a slight difference in behavior compared to upstream
- given that armeria's integration acts more as a client rather than a
proxy, the default virtual host will be used when the authority is null.

Modifications:

- Introduced `VirtualHostMatcher` which pre-computes a map for quicker
virtual host matching.
- `RouteEntry#index`, `VirtualHost#index` are exposed publicly. This
allows for easier referencing to the selected route.

Result:

- xDS clients route requests based on the authority

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->