Add `endResponseWithChild` to `RequestLogBuilder` (#6294)
## Motivation

With `endResponseWithLastChild`, `RequestLogBuilder` allows completing
the request log with the last child. However, there are situations where
we need to complete it with an arbitrary child log. This is the case,
for example, with request hedging (#5200), where multiple requests race
with each other, and any child log may be used as to complete the parent
log.

## Modifications

- Added `endResponseWithChild` to `RequestLogBuilder` to allow
completing a request log with an arbitrary child log.

There are no breaking changes and no behavior changes outside of the
newly added method.

## Result

Users are now able to complete a request log with an arbitrary child log
via `RequestLogBuilder.endResponseWithChild`.
This also unblocks #5200.

There are no breaking changes.