Add option to force reload values in `AsyncLoader` (#6328)
Motivation:

If `AsyncLoader` provides a way to load the cached value dynamically, we
may flexibly respond to external changes. For example, if the
certificates of Athenz changes, the cached OAuth2 keys need to be
refreshed.

Modifications:

- Main change) Add `load(boolean forceLoad)` to `AsyncLoader` to fetch a
new value from the given loader.
- Add `refreshAfterLoad()` option to `AsyncLoaderBuilder` to easily
refresh a cached value after the load.
- `Add `name()` option to `AsyncLoaderBuilder`.
  - The name is used to a prefix when logging.
- Cached values are excluded from the logging because they can be
sensitive when used with a token client.

Result:

You can now focibliy refresh the cached value in `AsyncLoader`