Implement unmasking for Thrift `FieldMasker` (#6330)
Motivation:

Following the implementation of
https://github.com/line/armeria/pull/6311, this changeset attempts to
implement deserialization of masked fields.

Doing so would allow users to define a `FieldMasker` as follows:
```
FieldMasker.builder()
  .addMasker(
    NormalFooStruct.class,
    struct -> encryptToString(struct),
    str -> return decryptToStruct(str));
```

Conceptually, this implementation is a corollary of
`MaskingBeanDeserializerModifier`.
There is no change in the public API as
`FieldMaskerBuilder#addMasker(Class, Function, Function)` is already
exposed.

Modifications:

- Added `UnMaskingTProtocol`, `TMaskingDeserializer`,
`UnMaskingContexts` which helps with the implementation of unmasking
- Deserialization requires instantiation of the default instance. To aid
with this, `TBaseCache` has been introduced.

Result:

- The unmasking API functions correctly for thrift structs