Fix the flaky `RetryingRpcClientTest.doNotRetryWhenResponseIsCancelled()`
Motivation:

https://github.com/line/armeria/actions/runs/16953427414/job/48050706187?pr=6353#step:10:2165
```java
RetryingRpcClientTest > doNotRetryWhenResponseIsCancelled() FAILED
    java.lang.AssertionError:
    Expecting actual not to be null
        at com.linecorp.armeria.it.client.retry.RetryingRpcClientTest.doNotRetryWhenResponseIsCancelled(RetryingRpcClientTest.java:353)
```
The flaky test assumes that the request is aborted by the response
cancellation before the request is sent to the wire. However, if
the connection is already established, the request might be written
before `response.cancel(true)` is invoked.

Modifications:

- Add a delay before writing the request so it can be cancelled earlier.

Result:

Less noisy CI builds.