Fix incorrect stream ID used for HTTP/2 flow control
Motivation:

Armeria internally maintains request IDs for a connection which
are monotonically increased starting from 1.
https://github.com/line/armeria/blob/603fc8685059ffafed5096261a14fd502f0be33f/core/src/main/java/com/linecorp/armeria/server/Http2RequestDecoder.java#L211
A stream ID for HTTP/2 flow control is calculated from the request ID
using (id * 2 + 1).
https://github.com/line/armeria/blob/603fc8685059ffafed5096261a14fd502f0be33f/core/src/main/java/com/linecorp/armeria/internal/common/InboundTrafficController.java#L150-L152
That formula must not be used on the server side because the client may
skip a stream ID or a request may be rejected before the ID is incremented.

Modifications:

- Use the actual stream ID provided by Netty instead of deriving from the
  request ID.

Result:

Fixed HTTP/2 flow control by using a correct stream ID for computing the
window size.