Fix a bug where closing a websocket session could send a `RST_STREAM` frame (#6375)
Motivation:

Currently, once an `HttpRequest` and `HttpResponse` is completed, the
underlying HTTP2 stream is cancelled using a `RST_STREAM`.
This makes sense for normal HTTP constructs since it indicates that we
are no longer interested in the request, and we would like to release
resources associated with it.

However, some protocols such as WebSockets implement their own graceful
shutdown procedure.

In detail, Armeria's `HttpRequest`, `HttpResponse` implements WebSocket
graceful shutdown and the reactive stream implementation is closed when
a `CLOSE` frame is both sent and received.
However, although the websocket session is completed, the underlying
HTTP2 stream may not necessarily be complete.
Protocol-wise, there is an inherent discrepancy between websocket
session completion and HTTP2 stream completion.

The current implementation defaults to sending a `RST_STREAM`
immediately once the corresponding `HttpRequest` and `HttpResponse`. For
websockets, I propose that a delay is given so that the remote has a
chance to end the stream.

Assuming that we will tie the lifecycle of inbound `WebSocket`s with
outbound `WebSocket`s (so sending a close frame also closes the inbound)
in #6357 , this option can be thought of similar to netty's
`forceCloseTimeoutMillis` option. (which acts as a timeout since sending
the CLOSE frame)

Modifications:

- Added a `closeHttp2StreamDelayMillis` option to `ServiceConfig` and
relevant implementations
- `WebSocketService` sets a `closeHttp2StreamDelayMillis` of 10 seconds
by default
- `HttpServerHandler` decides when to send a `RST_STREAM` based on the
`closeHttp2StreamDelayMillis`
- Added `Http2StreamLifecycleHandler` which maintains the lifecycle of
reset futures. This ensures that scheduled futures aren't leaked for
servers with high throughput.
- Every time a request/response is closed but the corresponding stream
is alive, `maybeResetStream` is called.
- Every time a stream is closed, `notifyStreamClosed` is called to clean
up possibly scheduled futures.

Result:

- Users can set a timeout for closing a websocket session
- Fix a bug where closing a websocket session could send a `RST_STREAM`
frame

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->