Introduce the `xds-api` module (#6403)
Motivation:

Envoy utilizes pgv to validate the schema of a resource.
ref:
https://www.envoyproxy.io/docs/envoy/latest/configuration/operations/config_validation/config_validation

Given that upstream `libs.controlplane.api` doesn't support this, I
propose that we publish our own jar containing APIs along with
validators.
See https://github.com/jrhee17/armeria/pull/43 for a sample run.

Going forward, this can be further developed so that we can add our own
annotations. This may be useful for marking which fields are supported
by Armeria - this can be further exposed to users via metrics/logs (in
case a control plane specifies fields not supported) and in the UI later
on.

Modifications:

- Workflows are added to periodically check if upstream envoy version is
updated
- `xds-apply-updates.yml`: Runs the update script and creates a pull
request
- `xds-compare-versions.yml`: Compares envoy versions with the current
version and determines whether an update is needed
- `xds-sync-apis.yml`: An umbrella task which invokes the above two
workflows periodically
- Scripts are added under `/xds-api/tools`
- `upstream-patch.sh`: Creates two worktrees corresponding to each envoy
version, and computes a patch. The patch is applied to the working
directory.
- `update-sha.sh`: Updates the sha of each version (envoy and
dependencies) and writes `API_SHAS` and `envoy_version` files. This has
been forked from upstream (which is also apache 2.0 licensed)
- `update-api.sh`: Updates the proto directory according to the
`API_SHAS` and `envoy_version` files. This has also been forked.
- Added dependencies to `pgv-java-stub` and `protoc-gen-validate`, each
of which generates validator files and applies validation at runtime.

Result:

- The `xds` module now relies on the `xds-api` module.

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->