Do not warn for certificates missing both SAN and CN (#6477)
Motivation:

Both CN and SAN would be absent if a certicate is public root CA.
Warning for CA is false positive and noisy.
In addition, since this is related to metrics, strict validation isn’t
necessary. If there’s an issue with the certificate, it will fail at the
TLS handshake level and the user will notice it.

Discord discussion:
https://discord.com/channels/1087271586832318494/1390342361770365020/1433043355566211254

Modifications:

- Change the log level in `CertificateUtil` to debug from warn.

Result:

You no longer see false positive warnings for certificate metrics.