Stop producing text content preview for non-text types with charset (#6485)
Motivation:

When no `ContentPreviewer` is selected to produce a content preview,
`TextContentPreviewer` acts as a fallback not only when the content type
is textual but also when a charset is specified. For example, a request
with a `Content-Type` header of `image/png; charset=utf-8` will still be
recorded because the charset is specified.

Modifications:

- Stop producing text content preview for non-text types with charset.

Result:

- **Breaking change**: non-textual content with charset is no longer
recorded by the `ContentPreviewer` fallback.