Apply `pgv` to `XdsBootstrap` resources (#6468)
Motivation:

Following the introduction of the `xds-api` module, we can now use `pgv`
to validate resources according to validations defined upstream.
As a bonus, this will be useful for the upcoming error-handling PR, as
it gives us a reliable way to test how invalid resources should notify
users.

Modifications:

- Applied pgv validation to each resource type
- Fixed failing tests so that they pass `pgv`

Result:

- proto-gen validation is applied to `XdsBootstrap`

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->