Introduce `XdsResource#version`, `XdsResource#revision` (#6470)
Motivation:

In order to gain insight into an xDS-based client, it is important to
know the version of the resource being used.
This PR attempts to introduce `XdsResource#version` which represents the
version sent by the server for a specific resource.

Additionally, `version` is represented as a string which may be
difficult to reason about or record as metrics efficiently. For this
reason, an additional `XdsResource#revision` field is introduced which
denotes how many times a `version` has been changed for a
ConfigSourceClient.

In order to denote which resources resulted in errors, a `resourceName`
field has been added to `SnapshotWatcher`. If a name is not found,

Refactoring) While implementing this approach, a bug was found where a
nonce is not reset per gRPC stream.
As such, SotwXdsStream has been refactored and a `ActualStream` has been
extracted which represents an actual gRPC stream. The nonce lives on the
`ActualStream`, whereas the version lives on the conceptual stream level
which is consistent with upstream implementation.

Additionally, a per-type resource cache has also been introduced to
cache resources immediately after a version has been incremented. This
is necessary if a resource is 1) quickly subscribed 2) unsubscribed 3)
and subscribed again.

Modifications:

- Introduced `version`/`revision` fields for each resource
- `SnapshotWatcher` now also notifies the resource name on error
- `onError` is also emitted for `sotw` type objects so that users can be
aware of these errors
- `SotwXdsStream` has been refactored so that it better follows the
behavior of upstream
- `VersionManager` now updates the version for the sotw stream and
manages the revision
- `SubscriberStorage#ResourceCache` caches resources for the most recent
version
- Misc) `XdsResource` now only outputs `(name, type, version, revision)`
by default for simplicity

Result:

- `XdsResource#version`, `XdsResource#revision` can be used to record
logs/metrics

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->