`XdsBootstrap` leaves a warning log on errors by default (#6489)
Motivation:

While `SnapshotWatcher`s provide a way for users to opt into logging
errors, this is not enabled by default.
Additionally, there is no way to notify users that secondary bootstrap
resources have not been loaded correctly.
For this reason, I propose that a default `SnapshotWatcher` be added to
a `XdsBootstrap` where the default watcher logs on errors. This watcher
is added to `AbstractRoot` and bootstrap constructs by default.

In order to ensure parsing failures are propagated,
`AbstractResourceNode#doOnChanged` now also propagates errors to the
parent on exception.

Note that transient errors (network errors while fetching resources) are
not logged at this level.

Modifications:

- A xDS-dedicated `EventExecutor` is used for the default event loop.
- `XdsBootstrapBuilder` is introduced, and users are able to set the
default `SnapshotWatcher`
- The default `SnapshotWatcher` is added to all bootstrap
constructs/`AbstractRoot` implementations
- If `AbstractResourceNode#doOnChanged` throws an exception, the error
is propagated to the parent node

Result:

- Parsing/configuration errors are propagated to the user-specified
`SnapshotWatcher`
- When an error occurs, a warning log is logged by default

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->