Users can easily create a certificate chain for testing using SignedCertificateExtension (#6519)
Motivation:

This is the first of a series of PRs included by
https://github.com/line/armeria/pull/6516.

Currently, our mTLS tests are not realistic in that:
- Certificate chains aren't verified
- We set TLS configurations to not verify the peer, which means in
reality there's no point in using mTLS.

For this, I propose that a `SignedCertificate` is introduced. This is a
certificate which is signed by another certificate (another
`SignedCertificate` or `SelfSignedCertificate`). A JUnit `Extension` has
also been introduced so users can easily test this behavior.

In the process, I found that `SelfSingedCertificateNameType` has a typo,
and isn't adding much value. This enum has been removed.

Modifications:

- Introduced `SignedCertificate`, `SignedCertificateExtension`
- `SelfSignedCertificate`, `SelfSignedCertificateExtension` inherits the
newly introduced constructs
- Default behavior has been modified so that all generated certificates
can act as a ca.
- `SelfSingedCertificateNameType` has been removed

Result:

- Users can easily create a certificate chain for testing using
`SignedCertificateExtension`

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->