Fix a bug where SNI doesn't reflect the `:authority` of a request (#6522)
Motivation:

This is part of an ongoing effort to apply per-request TLS:
https://github.com/line/armeria/pull/6516

Currently, SNI is determined as follows:
- If an `Endpoint` has a hostname, the hostname is used
- Otherwise, the `:authority` at the time of setting the `Endpoint` is
used


https://github.com/line/armeria/blob/ee1836f9de74a8c10166860f533d08cfe4930079/core/src/main/java/com/linecorp/armeria/internal/client/DefaultClientRequestContext.java#L584-L593

When using the `:authority`, it does not consider headers that might
have been set after the endpoint is determined (e.g. in the decorators)

I propose that the final SNI is determined right before `PoolKey`
generation. This will be useful in the upcoming PR to determine the
exact `TLS` specification associated with a HTTP request.

Modifications:

- Moved SNI determining logic to `HttpClientDelegate`
- Removed unnecessary recomputation of the `Origin` header

Result:

- More consistent behavior regarding SNI determination

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->