Introduce `ConfigSourceLifecycleObserver` (#6488)
Motivation:

This PR introduces `ConfigSourceLifecycleObserver` which observes a
`ConfigSourceClient`.
This may be useful when users would like to monitor the client which
fetches xDS resources and can be as an equivalent of subscription
statistics in envoy
ref:
https://www.envoyproxy.io/docs/envoy/latest/configuration/overview/mgmt_server#xds-subscription-statistics

In order to allow users to select the `MeterRegistry`/`MeterIdPrefix` a
`XdsBootstrapBuilder` is introduced. While it is trivial to allow users
to customize the `ConfigSourceLifecycleObserver`, I don't think it's
necessary at this stage and could be easily done later on if there is
need for it.

Modifications:

- Added a `ConfigSourceLifecycleObserver` and a default implementation
`DefaultConfigSourceLifecycleObserver`
- `ConfigSourceLifecycleObserver` callback methods are called from
`SotwXdsStream`
- Added a `XdsBootstrapBuilder` which allows users to customize the
`MeterRegistry`, `MeterIdPrefix`
- Modified so that the default `EventExecutor` used is dedicated for
pushing xDS updates.

Result:

- Users can observe the behavior of `ConfigSourceClient` via metrics

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->