Disallow query strings in a route path
Motivation:

The query string in a path of `Route` is treated as a part of routing
path. However, when a service is actually resolved, the query string of
the request path is excluded from the target path. As a result, the
router fails to locate the service.
https://github.com/line/armeria/blob/95ed9e1cfb075da4f04d7e5d5fa5adc63627282d/core/src/main/java/com/linecorp/armeria/server/RoutingContext.java#L82-L88

Modifications:

- Raise `IllegalArgumentException` if the path of `route()` contains
  query strings and guide users to use `RouteBuilder.matchParams()`

Result:

`Route` now properly rejects paths with query strings to avoid incorrect
routing behavior.