Introduce `ServerTlsSpec` (#6550)
Motivation:

In an effort to generalize usage of the `TlsSpec` abstraction introduced
in https://github.com/line/armeria/pull/6542, I propose that a new
`ServerTlsSpec` is introduced.

Note that this PR is a refactoring change which doesn't introduce new
features, but organizes the previous logic so that TLS-related logic is
applied consistently.

Modifications:

- `ServerTlsSpec`, `ServerTlsSpecBuilder` is introduced
- `ClientSslContextFactory` is removed, and `SslContextFactory` is used
as the main factory to convert `TlsSpec` to `SslContext`.
- `IgnoreHostsTrustManager` is now removed as the functionality is
replaced with `TlsPeerVerifier`s
- `SslContextUtil#createSslContext` is removed and replaced with
`SslContextUtil#toSslContext`
- A single `SslContextFactory` is instantiated for a single server
- `SslContextFactory` acts as the main source of recording metrics,
hence `Server`-specific metric logic has been removed
- `TlsProviderMapping` has been modified to use `SslContextFactory` to
maintain `SslContext`s
- The previous logic using `SslContextBuilder` has been removed and
abstracted to a `VirtualHostTlsSetter`
- `VirtualHostTlsSetter` sets TLS related parameters, and generates a
`ServerTlsSpec`

Result:

- `SslContextFactory`, `ServerTlsSpec` is used to create `SslContext`s
for server-side.
<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->