Add ping-related events to `ConnectionPoolListener` (#6552)
Motivation:

We currently lack observability into `KeepAliveHandler`.
This PR attempts to expose events so that users can be aware of when
pings are sent/received for the client-side.
Additionally, a `CloseHint` is added to the `connectionClosed` interface
so users can better understand why a connection has been closed.

In order to better define the behavior when `ConnectionPoolListener` is
invoked, a `ConnectionEventListener` is introduced.
This class is attached to the `Channel` for bookkeeping, and eventually
delegating events to the `ConnectionPoolListener`.

The current `MetricCollectingConnectionPoolListener` hasn't been
modified for this iteration as it is complex, and removing the purging
logic is probably better than applying changes in the current PR.
I don't think handling this is necessary in this PR, but can do so if
other maintainers feel it is useful.

Modifications:

- Added `ConnectionPoolListener#pingWrite`,
`ConnectionPoolListener#pingAck`, and a variant of
`ConnectionPoolListener#connectionClosed` which exposes a `closeHint`
  - `pingWrite` is invoked every time a ping is written
  - `pingAck` is invoked every time a written ping is acknowledged
- `onPingAck()` represents an acknowledgement of a ping. Hence,
`ClientHttp1KeepAliveHandler` has been modified to be able to handle
`onPingAck()`.
- A `ConnectionEventListener` is attached to a `Channel` on bootstrap
registration.
- Lifecycle methods are invoked for the `ConnectionEventListener`, which
in turn delegates these to the `ConnectionPoolListener`
- `KeepAliveHandler` maintains a reference to a
`ConnectionEventListener`, and notifies events to the listener.

Result:

- Closes #6539

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->